requirements.txt
numpy==1.26.4
viam-sdk==0.25.2
matplotlib
pyinstaller
torchvision
torch
pytest
pillow

